Update to query expression
Any of the queries that use anything other than LIKE were not working for me. The percent placeholders were just empty strings instead of adding quotes around the value to be queried. SO it was attr_name = value SO I changed it to give you this result: attr_name = 'value'    it can't evaluate the query if it doesn't know that is a string :)